backend/pyproject.toml
[project]
name = "direkt-laminat-backend"
version = "0.1.0"
description = "PDF Order Pipeline for DirektLaminat — AI-powered PDF extraction to Monitor ERP XML"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.14.0",
    "openai>=1.50.0",
    "PyMuPDF>=1.24.0",
    "python-multipart>=0.0.12",
    "lxml>=5.3.0",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "python-jose[cryptography]>=3.3.0",
    "asyncpg>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "DTZ", "T20", "ICN", "PIE", "PT", "RSE", "RET", "SIM", "TID", "TCH", "ARG", "ERA", "PGH", "PL", "RUF"]
ignore = ["S101", "PLR2004"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
